Program works for doubles now

+++ src/embs/Edge.java
@@ -2,10 +2,10 @@ package embs;
 
 public class Edge {
 
-  public final int weight;
+  public final double weight;
 
 
-  public Edge(int weight) {
+  public Edge(double weight) {
     this.weight = weight;
   }
